requirements.txt
langchain-astradb
langchain 
langchain-openai
datasets
pypdf
python-dotenv
google-generativeai
langchain-community
langchain-google-genai
streamlit

-e .